linker issue unresolved

+++ Header Files/physics.h
@@ -5,6 +5,6 @@
 #include <chrono>
 
 float gravity(float v_velocity, float timestep);
-std::vector<int> updatePos(std::vector<float> position, float v_velocity, float h_velocity, float timestep);
+std::vector<float> updatePos(std::vector<float> position, float v_velocity, float h_velocity, float timestep);
 
 #endif
